requirements.txt
requests==2.28.2
python-socketio==5.8.0
web3==5.31.3
websocket-client==1.3.2
ccxt==3.0.51
